fixed directory structure
