--- token/transfer-hook-example/transfer-hook-example/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "spl-transfer-hook-example"
-version = "0.1.0"
-description = "Solana Program Library Transfer Hook Example Program"
-authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
-repository = "https://github.com/solana-labs/solana-program-library"
-license = "Apache-2.0"
-edition = "2021"
-
-[features]
-no-entrypoint = []
-test-sbf = []
-
-[dependencies]
-arrayref = "0.3.6"
-solana-program = "1.14.12"
-spl-tlv-account-resolution = { version = "0.1.0" , path = "../../libraries/tlv-account-resolution" }
-spl-transfer-hook-interface = { version = "0.1.0" , path = "../transfer-hook-interface" }
-spl-type-length-value = { version = "0.1.0" , path = "../../libraries/type-length-value" }
-
-[dev-dependencies]
-solana-program-test = "1.14.12"
-solana-sdk = "1.14.12"
-spl-token-2022 = { version = "0.6",  path = "../program-2022", features = ["no-entrypoint"] }
-spl-token-client = { version = "0.5",  path = "../client" }
-
-[lib]
-crate-type = ["cdylib", "lib"]
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
