+++ token/transfer-hook-interface/Cargo.toml
@@ -7,16 +7,12 @@ repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
 edition = "2021"
 
-[features]
-offchain-client = ["dep:solana-sdk"]
-
 [dependencies]
 arrayref = "0.3.6"
 bytemuck = { version = "1.13.0", features = ["derive"] }
 num-derive = "0.3"
 num-traits = "0.2"
-num_enum = "0.6.1"
-solana-sdk = { version = "1.14.12", optional = true }
+num_enum = "0.5.9"
 solana-program = "1.14.12"
 spl-tlv-account-resolution = { version = "0.1.0" , path = "../../libraries/tlv-account-resolution" }
 spl-type-length-value = { version = "0.1.0" , path = "../../libraries/type-length-value" }
