+++ token/transfer-hook-interface/src/lib.rs
@@ -10,7 +10,7 @@
 pub mod error;
 pub mod instruction;
 pub mod invoke;
-#[cfg(feature = "offchain-client")]
+pub mod manager;
 pub mod offchain;
 
 // Export current sdk types for downstream users building with a different sdk version
