--- token/transfer-hook-interface/transfer-hook-interface/src/lib.rs
@@ -1,53 +1,0 @@
-//! Crate defining an interface for performing a hook on transfer, where the
-//! token program calls into a separate program with additional accounts after
-//! all other logic, to be sure that a transfer has accomplished all required
-//! preconditions.
-
-#![allow(clippy::integer_arithmetic)]
-#![deny(missing_docs)]
-#![cfg_attr(not(test), forbid(unsafe_code))]
-
-pub mod error;
-pub mod instruction;
-pub mod invoke;
-pub mod manager;
-pub mod offchain;
-
-// Export current sdk types for downstream users building with a different sdk version
-pub use solana_program;
-use solana_program::pubkey::Pubkey;
-
-/// Namespace for all programs implementing transfer-hook
-pub const NAMESPACE: &str = "spl-transfer-hook-interface";
-
-/// Seed for the state
-const EXTRA_ACCOUNT_METAS_SEED: &[u8] = b"extra-account-metas";
-
-/// Get the state address PDA
-pub fn get_extra_account_metas_address(mint: &Pubkey, program_id: &Pubkey) -> Pubkey {
-    get_extra_account_metas_address_and_bump_seed(mint, program_id).0
-}
-
-/// Function used by programs implementing the interface, when creating the PDA,
-/// to also get the bump seed
-pub fn get_extra_account_metas_address_and_bump_seed(
-    mint: &Pubkey,
-    program_id: &Pubkey,
-) -> (Pubkey, u8) {
-    Pubkey::find_program_address(&collect_extra_account_metas_seeds(mint), program_id)
-}
-
-/// Function used by programs implementing the interface, when creating the PDA,
-/// to get all of the PDA seeds
-pub fn collect_extra_account_metas_seeds(mint: &Pubkey) -> [&[u8]; 2] {
-    [EXTRA_ACCOUNT_METAS_SEED, mint.as_ref()]
-}
-
-/// Function used by programs implementing the interface, when creating the PDA,
-/// to sign for the PDA
-pub fn collect_extra_account_metas_signer_seeds<'a>(
-    mint: &'a Pubkey,
-    bump_seed: &'a [u8],
-) -> [&'a [u8]; 3] {
-    [EXTRA_ACCOUNT_METAS_SEED, mint.as_ref(), bump_seed]
-}
